Fix project dependencies
Fix issue with project internal dependencies
Refactore rprpp.cpp file
