+++ apps/glfwapp/CMakeLists.txt
@@ -11,15 +11,10 @@ set(GLFW_INSTALL OFF CACHE INTERNAL "Glfw install")
 FetchContent_MakeAvailable(glfw)
 
 set(HEADERS
-    ../common/rpr_helper.h
-    ../common/HybridProRenderer.h
-    ../common/RprPostProcessing.h
     Dx11App.h
 )
 
 set(SOURCES
-    ../common/HybridProRenderer.cpp
-    ../common/rpr_helper.cpp
     Dx11App.cpp
 )
 add_executable(glfwapp ${SOURCES} ${HEADERS} main.cpp)
@@ -30,12 +25,7 @@ _copy_file_to_target(glfwapp ${RPRSDK_DIR}/binWin64/RadeonProRender64.dll)
 _copy_file_to_target(glfwapp ${CMAKE_CURRENT_SOURCE_DIR}/../assets/teapot.obj)
 _copy_file_to_target(glfwapp ${CMAKE_CURRENT_SOURCE_DIR}/../assets/envLightImage.exr)
 
-target_include_directories(glfwapp
-    PRIVATE ${RPRSDK_DIR}/inc
-)
-
 target_link_libraries(glfwapp 
     glfw 
-    rprpp
-    ${RPRSDK_DIR}/libWin64/RadeonProRender64.lib
+    appcommon
 )
